Remove length parameters from usb get_string functions.
With PyUSB beta version 1.0.0b2 the string length paramtere in 
util.get_string got removed and the function just fetches up to the 
maximum possible string length (126 UTF-16 characters). For 
compatibility with earlier alpha and beta releases we call first the
new version w/o length and fall back to the previous if it fails.