Added Type 3 Tag specific flags to examples/tagtool.py tag emulation
Similar flags as for type 3 tag formatting needed for emulation of debug
tags. All aspects of the type 3 attribute block can be manually set to
whatever valid or invalid combinations. The default is still to do "the
right thing". Also corrected handling of tag size: The --size parameter 
is now adjusted to always fit the NDEF data, it's now the minimum size 
of the NDEF data area.