Correct formatting of FeliCa Lite-S tags
When determining the number of user data blocks by forward reading until
an exception occurs, the last data block of a FeliCa Lite and Lite-S tag
is in fact not usable.
  
A FeliCa Lite / Lite-S tag does not listen to system code 12FC when it 
comes out of factory. To enable 12FC we must set byte 3 of the memory 
configuration block 0x88 to 0x01.