Use PN53x InCommunicateThru for tag communication (except for Jewel/Topaz)
The PN53x chips limit the the range of command codes which may be send to
a Mifare Ultralight (Type 2 Tag) card when using the PN53x InDataExchange
method. For newer versions such as the Ultralight EV1 or NTAGs this meant
some card commands could not be executed. Using InCommunicateThru solved
that issue but now requires the driver to memorize the modulation type
(actually it's the different frame format between Mifare and Jewel) to 
switch to InDataExchange for a Jewel/Topaz card. For a type 4 tag the 
driver now activates the ISO-DEP implementation in tags/tt4.py by no 
longer announcing the ISO-DEP support of the PN53x.