Disable CRC check for type 2 tag responses also on PN53x.
With late discovery that the PN53x chips allow access to the Contactless
Interface Unit (CIU) registers and the CIU registers are pretty similar
to PN512 (for which documentation is available), it is now possible to 
disable the CRC-A check for type 2 tags in the same way as for the Sony
RC-S380. Doing the CRC check in software means that the PN53x will also 
pass ACK and NAK responses to nfcpy, this in turn means that we can use
InCommunicateThru to avaoid the command code restrictions that are baked
into InDataExchange.