Read page by page for type 2 tag memory dumps.
Type 2 tags may be read protected from a page that is not a multiple 
of 4. Because the read command returns 16 byte (4 pages) the previous
strategy was to read and cache pages aligned at multiples of 4. The
corrected but unfortunately also slower strategy is to read page by page
and only use the first 4 bytes. And because memory rolls over at a
protection boundary it is also not easy to determine when and how many 
of the trailing bytes would be valid (although it would be possible to
read until a protected page is hit and then jump back 3 pages and read
single increments, but this is ugly and there's also anticollision
needed whenever a read has failed.