Improve format method
Cover the case that absolutely no block can be read (quite esotericc
case but least theoretically this may happen). Also the tests for how
many blocks can be read or written in a single command are now
independent of the total number of blocks (Nbr and Nbw may be larger
than NmaxB).